pyproject.toml
[tool.poetry]
name = "metadata"
version = "0.1.0"
description = "Create and update file metadata"
authors = ["Indoc Research"]

[tool.poetry.dependencies]
fastapi = "0.63.0"
fastapi-health = "0.4.0"
fastapi-sqlalchemy = "0.2.1"
fastapi-utils = "0.2.1"
pilot-platform-common = "^0.0.25"
psycopg2-binary = "2.9.3"
python = "^3.9"
requests = "2.24.0"
sqlalchemy-utils = "0.38.2"
uvicorn = "0.17.5"
python-json-logger = "0.1.11"
strawberry-graphql = {version = "0.112.0", extras = ["extras"]}

[tool.poetry.dev-dependencies]
pytest = "6.2.4"
coverage = "5.3"
flake8 = "4.0.1"
alembic = "^1.7.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
